+++ packages/frontend/package.json
@@ -60,9 +60,9 @@
     "@ember/optional-features": "^1.1.0",
     "@ember/render-modifiers": "^1.0.2",
     "@ember/test-helpers": "^1.7.1",
-    "@embroider/compat": "^0.9.0",
-    "@embroider/core": "^0.9.0",
-    "@embroider/webpack": "^0.9.0",
+    "@embroider/compat": "^0.10.0",
+    "@embroider/core": "^0.10.0",
+    "@embroider/webpack": "^0.10.0",
     "@fortawesome/ember-fontawesome": "^0.2.1",
     "@fortawesome/free-brands-svg-icons": "^5.12.0",
     "@fortawesome/free-solid-svg-icons": "^5.12.0",
@@ -115,7 +115,7 @@
     "ember-cli-page-object": "^1.16.1",
     "ember-cli-postcss": "^5.0.0",
     "ember-cli-sri": "^2.1.1",
-    "ember-cli-typescript": "^3.1.1",
+    "ember-cli-typescript": "^3.1.2",
     "ember-cli-uglify": "^3.0.0",
     "ember-component-css": "^0.7.4",
     "ember-computed-promise-monitor": "^0.4.0",
@@ -154,9 +154,9 @@
     "ember-web-app": "^3.0.1",
     "eslint": "^6.7.2",
     "eslint-config-prettier": "^6.7.0",
-    "eslint-plugin-ember": "7.7.1",
+    "eslint-plugin-ember": "7.7.2",
     "eslint-plugin-node": "^10.0.0",
-    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-qunit": "^4.0.0",
     "git-rev-sync": "^1.12.0",
     "loader.js": "^4.7.0",
@@ -167,7 +167,7 @@
     "qunit-assertions-extra": "^0.8.1",
     "qunit-console-grouper": "^0.2.0",
     "qunit-dom": "^0.9.2",
-    "rollup": "^1.27.11",
+    "rollup": "^1.27.13",
     "rollup-plugin-babel": "^4.3.3",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-filesize": "^6.2.1",
