+++ packages/frontend/tests/helpers/setup-workers.ts
@@ -1,6 +1,8 @@
 import { TestContext } from 'ember-test-helpers';
 import WorkersService, { CRYPTO_PATH } from 'emberclear/services/workers';
-import { handleMessage, CryptoMessage } from 'emberclear/workers/crypto';
+import { handleMessage } from 'emberclear/workers/crypto/messages';
+
+type CryptoMessage = import('emberclear/workers/crypto/messages').CryptoMessage;
 
 /**
  *
